handdle values function
